requirements.txt
asana==0.9.1
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
gunicorn==20.0.0
idna==2.8
oauthlib==3.1.0
psycopg2==2.8.4
python-decouple==3.3
pytz==2019.3
requests==2.22.0
requests-oauthlib==0.8.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
